% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize string}
\usage{
capitalize(string)
}
\arguments{
\item{string}{A string that you want to capitalize.}
}
\description{
Capitalize first letter in the string.
}
