% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_playlist.R
\docType{data}
\name{clean_playlist}
\alias{clean_playlist}
\title{Clean playlist}
\format{
A data table with 159 rows and 3 variables:
\describe{
  \item{track_name}{Track name}
  \item{artist_name}{Artist name}
  \item{playlist_name.playlists.name}{Playlist name}
}
}
\usage{
clean_playlist
}
\description{
A dataset, selected from Wiktor's data from spotify,result of using prepare_playlists()
on JSON file, prepared for testing and vignettes
}
\keyword{datasets}
