% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_stream_his2.R
\docType{data}
\name{clean_stream_his2}
\alias{clean_stream_his2}
\title{Clean streaming history}
\format{
A data table with 100 rows and 7 variables:
}
\usage{
clean_stream_his2
}
\description{
Second dataset with prepared streaming history,result of using make_streaming_history_dt() on JSON file,
selected from Wiktor's data from spotify,
prepared for testing and vignettes.
}
\keyword{datasets}
