% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_history.R
\name{compare_history}
\alias{compare_history}
\title{Common artists / artists and tracks for two users.}
\usage{
compare_history(streaming_history_1, streaming_history_2, by_track = F)
}
\arguments{
\item{streaming_history_1}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it from user 1.}

\item{streaming_history_2}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it from user 2.}

\item{by_track}{Logical value, whether comparison should include tracks. Default is FALSE.}
}
\description{
Shows the same artists/ artists and tracks from history of two different users.
}
