% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_streaming_history.R
\name{complete_streaming_history}
\alias{complete_streaming_history}
\title{Make and prepare the streaming history data table}
\usage{
complete_streaming_history(folder_path)
}
\arguments{
\item{folder_path}{Path to a folder which contains file or files with 
Streaming History in json format. File(s) must contain "StreamingHistory" in the name.}
}
\value{
A data table containing streaming history from spotify suited for being used 
in rest of the functions from the package.
}
\description{
Make data table with make_streaming_history_dt and do prepare_streaming_history with, that
data table
}
