% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_listening.R
\name{continuous_listening}
\alias{continuous_listening}
\title{Measure the number of continuous listening.}
\usage{
continuous_listening(streaming_history, mins)
}
\arguments{
\item{streaming_history}{A data table containg streaming history from spotify.}

\item{mins}{Number of minutes which determine distance between listenings.}
}
\value{
A data table containg streaming history with additional column about listening number.
}
\description{
Listening number informs about number of songs listened in a row.
If distance between beginning of a song and end of a previous one is less than mins then listening number stays.
Otherwise increases. Counting begins from the newest song in streaming history.
}
