% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track_count.R
\name{count_tracks}
\alias{count_tracks}
\title{vizualize number of tracks that were played or skipped}
\usage{
count_tracks(streaming_history, only_skipped = FALSE, by = "year")
}
\arguments{
\item{streaming_history}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it.}

\item{only_skipped}{A logical scalar indicating whether to show all played tracks or only those that were skipped.}

\item{by}{A character vector indicating for which periods to split the time for counting tracks played/skipped. 
Permitted values are: "year", "month", "week", "day".}
}
\description{
vizualizes, using a bar chart, how many tracks were played or how many were skipped, each period (day, week, month, year).
}
