% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty_stream_his.R
\docType{data}
\name{dirty_stream_his}
\alias{dirty_stream_his}
\title{Dirty streaming history}
\format{
A data table with 100 rows and 4 variables:
}
\usage{
dirty_stream_his
}
\description{
A dataset with dirty streaming history, result of using make_streaming_history_dt() on JSON file,
 selected from Wiktor's data from spotify, prepared for testing and vignettes.
}
\keyword{datasets}
