% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_search_queries.R
\name{filter_search_queries}
\alias{filter_search_queries}
\title{Filter spotify searh queries}
\usage{
filter_search_queries(search_queries, start_date, end_date)
}
\arguments{
\item{search_queries}{}

\item{start_date}{}

\item{end_date}{}
}
\description{
Filter spotify search queries data table to contain only entries from specified time period
}
\seealso{
\code{\link{filter_streaming_history}}
}
