% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sec_to_hms.R
\name{from_sec_to_hms}
\alias{from_sec_to_hms}
\title{Change format to hours minutes seconds}
\usage{
from_sec_to_hms(x)
}
\arguments{
\item{x}{An integer vector. Number of seconds}
}
\value{
A character vector with format "X h Y min Z s", where Y < 60 and Z < 60
}
\description{
Change format form number of seconds to number of hours, minutes and seconds
}
\examples{
from_sec_to_min_sec(60)
from_sec_to_min_sec(134)
}
