% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{from_sec_to_hms_single}
\alias{from_sec_to_hms_single}
\title{from_sec_to_hms helper function.}
\usage{
from_sec_to_hms_single(x)
}
\arguments{
\item{x}{An integer vector. Number of seconds.}
}
\description{
Function to be applied to an integer, representing number of seconds,
 changes the format to "X h Y min Z s". Applied to each element of the vector by
  \code{from_sec_to_hms}.
}
