% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_many_skipped.R
\name{how_many_skipped}
\alias{how_many_skipped}
\title{Count skipped songs in given time period}
\usage{
how_many_skipped(
  streaming_history,
  start_date,
  end_date,
  as_percentage = FALSE
)
}
\arguments{
\item{streaming_history}{data.table containing streaming history, after 'prepare_streaming_history' was used on it}

\item{start_date}{POSIXt indicating start of the period of time to count skipped songs}

\item{end_date}{POSIXt indicating end of the period of time to count skipped songs}

\item{as_percentage}{logical. if FALSE (default) a number of skipped songs is returned,
otherwise a character vector indicating percentage of skipped songs in given period}
}
\description{
Counts how many songs were skipped in given time period. Returns either a number or a percentage of all songs
in given time period depending on user
}
