% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longest_session.R
\name{longest_session}
\alias{longest_session}
\title{Longest session info}
\usage{
longest_session(streaming_history, mins)
}
\arguments{
\item{streaming_history}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it.}

\item{mins}{Number of minutes which determine distance between listenings.}
}
\value{
A data table containg information about longest session.
}
\description{
Shows all information about longest session i.e. tracks, artists etc. Excludes skipped tracks.
}
