% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_search_queries_dt.R
\name{make_search_queries_dt}
\alias{make_search_queries_dt}
\title{Make data table with search queries}
\usage{
make_search_queries_dt(folder_path)
}
\arguments{
\item{folder_path}{Path to a folder which contains file with 
search queries in json format. File must contain 'SearchQueries" in the name.}
}
\value{
A data table with 3 columns(date, platformm, contry) 
containing details of Search Queries from spotify.
}
\description{
Make data table with search queries from Spotify data.
}
