% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_session_stats.R
\name{make_session_stats}
\alias{make_session_stats}
\title{Make session stats table}
\usage{
make_session_stats(session_dt)
}
\arguments{
\item{session_dt}{A data table containing information about listening session. 
Must have columns "artist_name", "track_name", "s_played", "start_time".}
}
\description{
Make a table containing stats about session - play time , number of tracks played, start and end.
}
