% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_streaming_history_dt.R
\name{make_streaming_history_dt}
\alias{make_streaming_history_dt}
\title{Make data table with streaming history}
\usage{
make_streaming_history_dt(folder_path)
}
\arguments{
\item{folder_path}{Path to a folder which contains file or files with 
Streaming History in json format. File(s) must contain "StreamingHistory" in the name.}
}
\value{
A data table with 4 columns(endTime, artistName, trackName, msPlayed) of characters 
containing details of Streaming History from spotify.
}
\description{
Make data table with streaming history from Spotify data.
}
