% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_played.R
\name{most_played}
\alias{most_played}
\title{Most played tracks/artist}
\usage{
most_played(
  streaming_history,
  start_date,
  end_date,
  track_or_artist,
  how_many = 10,
  show_skipped = TRUE
)
}
\arguments{
\item{streaming_history}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it.}

\item{start_date}{A POSIXt indicating start of the period of time.}

\item{end_date}{A POSIXt indicating end of the period of time.}

\item{track_or_artist}{A character string that specifies wheter to show most frequently played artists or tracks. 
Must be either "track" or "artist"}

\item{how_many}{A positive integer indicating how many most frequent tracks/artist to show, defoults to 10.}

\item{show_skipped}{A logic scalar indicating whether to include tracks that were skipped in counting.}
}
\description{
Shows most frequently played artist or track in given time period and counts how many times were they played.
}
