% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_tracks_by_hour.R
\name{number_of_tracks_by_hour}
\alias{number_of_tracks_by_hour}
\title{Visualize number of tracks played by hour}
\usage{
number_of_tracks_by_hour(
  streaming_history,
  start_date,
  end_date,
  by_weekday = FALSE,
  show_skipped = FALSE
)
}
\arguments{
\item{streaming_history}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it.}

\item{start_date}{A POSIXt indicating start of the period of time.}

\item{end_date}{A POSIXt indicating end of the period of time.}

\item{by_weekday}{A logic scalar indicating whether to show data for each weekday separately on the graph.}

\item{show_skipped}{A logic scalar indicating whether to include tracks that were skipped in counting.}
}
\description{
Vizualizes, using a bar chart, the number of tracks played on different hours. 
Either collectively or separetely for each day of the week. Can be set to inculde songs that were counted as skipped.
}
