% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_tracks_by_weekday.R
\name{number_of_tracks_by_period}
\alias{number_of_tracks_by_period}
\title{Visualize number of tracks played by hour}
\usage{
number_of_tracks_by_period(
  filtered,
  as_percentage = TRUE,
  include_skipped = FALSE
)
}
\arguments{
\item{filtered}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it.}

\item{as_percentage}{A logical scalar. If TRUE (dafault) bars show percentage of of all tracks played for each weekday.}

\item{include_skipped}{A logical scalar indicating whether to include tracks that were skipped}

\item{period}{A character vector indicating for which periods to split the time for counting tracks played}
}
\description{
Vizualizes, using a bar chart, the number of songs that were played on each of the weekdays in given time period. 
Can be set to inculde songs that were counted as skipped
}
