% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_tracks_by_weekday.R
\name{number_of_tracks_by_weekday}
\alias{number_of_tracks_by_weekday}
\title{Visualize number of tracks played by hour}
\usage{
number_of_tracks_by_weekday(
  streaming_history,
  start_date,
  end_date,
  show_skipped = FALSE,
  as_percentage = TRUE
)
}
\arguments{
\item{streaming_history}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it.}

\item{start_date}{A POSIXt indicating start of the period of time.}

\item{end_date}{A POSIXt indicating end of the period of time.}

\item{show_skipped}{A logical scalar indicating whether to include tracks that were skipped in counting.}

\item{as_percentage}{A logical scalar. If TRUE (dafault) bars show percentage of of all tracks played for each weekday.}
}
\description{
Vizualizes, using a bar chart, the number of songs that were played on each of the weekdays in given time period. 
Can be set to inculde songs that were counted as skipped
}
