% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_day_comparison.R
\name{plot_day_comparison}
\alias{plot_day_comparison}
\title{Compare two users day play time}
\usage{
plot_day_comparison(two_users_dt, user1_name = "User 1", user2_name = "User 2")
}
\arguments{
\item{two_users_dt}{A data table with two users play times on a specific day, made with \code{\link{make_two_users_dt}}}

\item{user1_name}{A character vector, name to display for the first user.}

\item{user2_name}{A character vector, name to display for the second user.}
}
\description{
Compare on which times two users were listening to spotify during a specified day.
}
