% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_in_playlists_count.R
\name{plot_in_playlists_count}
\alias{plot_in_playlists_count}
\title{Visualize how much of music listened to was from playlists}
\usage{
plot_in_playlists_count(
  str_his_with_playlist_wide,
  as_percentage = FALSE,
  time_or_count = "time",
  time_unit = "minutes"
)
}
\arguments{
\item{str_his_with_playlist_wide}{A data table containing streaming history 
with information about playlists, in wide form, made with \code{\link{str_his_with_playlists_wide}}.}

\item{as_percentage}{A logical scalar indicating whether to show a percentage or an absolute value.}

\item{time_or_count}{A character vector indicating whether to count
number of tracks played or the time they were played.}

\item{time_unit}{A character vector indicating the time unit to use 
if showing an absolute value of time played. Permitted vales are "seconds", "minutes", "hours".}
}
\description{
Visualize, using a bar chart, how much music listened to was from playlists. 
Either counting the number of tracks played or play time.
}
