% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_playlist_popularity.R
\name{plot_playlist_popularity}
\alias{plot_playlist_popularity}
\title{Visualize popularity of different playlist}
\usage{
plot_playlist_popularity(
  str_his_with_playlists_long,
  time_or_count = "time",
  time_unit = "hours"
)
}
\arguments{
\item{str_his_with_playlists_long}{A data table containing streaming history with information about playlists,
in long form, made with \code{\link{str_his_with_playlists_long}}.}

\item{time_or_count}{A character vector indicating whether to count number of tracks played
or the time they were played.}

\item{time_unit}{A character vector indicating the time unit to use if showing an absolute value of time played.
Permitted vales are "seconds", "minutes", "hours".}
}
\description{
Visualize, using a bar chart, how popular were different playlists. Either by number of tracks played or play time.
}
