% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_searches.R
\name{plot_searches}
\alias{plot_searches}
\title{Visualize search queries}
\usage{
plot_searches(search_queries, additional = "country")
}
\arguments{
\item{search_queries}{A data table containing spotify search queries, made with \code{\link{make_search_queries_dt}}}

\item{additional}{A character vector, additional information to be visualized, can be either "country" or "platform".}
}
\description{
Visualize search queries
}
