% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_track_count.R
\name{plot_track_count}
\alias{plot_track_count}
\title{visualize number of tracks that were played or skipped}
\usage{
plot_track_count(streaming_history, only_skipped = FALSE, by = "year")
}
\arguments{
\item{streaming_history}{A data.table containing streaming history, 
after 'prepare_streaming_history' was used on it.}

\item{only_skipped}{A logical scalar indicating whether to show 
all played tracks or only those that were skipped.}

\item{by}{A character vector indicating for which periods 
to split the time for counting tracks played/skipped.
Permitted values are: "year", "month", "week", "day".}
}
\description{
visualizes, using a bar chart, how many tracks were played or how many were skipped, each period (day, week, month, year).
}
