% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_playlists.R
\name{prepare_playlists}
\alias{prepare_playlists}
\title{Preapare playlists data table}
\usage{
prepare_playlists(folder_path)
}
\arguments{
\item{folder_path}{Path to a folder which contains file with Playlists in json format. File must contain "Playlist" in the name.}
}
\value{
A data table containing the names of playlists and songs on them from spotify.
}
\description{
Preapare playlists data table
}
