% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_streaming_history.R
\name{prepare_streaming_history}
\alias{prepare_streaming_history}
\title{Prepares the streaming history data table}
\usage{
prepare_streaming_history(streaming_history)
}
\arguments{
\item{streaming_history}{A raw data table containing streaming history from spotify.}
}
\value{
A data table containg streaming history from spotify suited for being used in rest of the functions from the package.
}
\description{
Prepares the data table containig streaming history for futher oprations.
 Changes the names of the base columns to: end_time, artist_name, track_name, s_played. Changes the end_time 
 column from charcter to POSTIX date, s_played to a duration in seconds. Adds new columns: 
 start_time - estimated start time calculated from end_time and s_played, skipped - TRUE if the track was 
 played for less than 10s, weekday - weekdays derived from start_time date.  \cr
}
