% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_for_view.R
\name{session_for_view}
\alias{session_for_view}
\title{Change session data table for viewing}
\usage{
session_for_view(session_dt)
}
\arguments{
\item{session_dt}{A data table containing information about listening session. 
Must have columns "artist_name", "track_name", "s_played", "start_time".}
}
\description{
Changes the data table containing informations about a session (for example  form 
\code{\link{longest_session}}) to be in a format
   good for reading - spaces instead of "_" in column names, time of playing in 
   "Xh Ymin Zs" format, and have only columns with artist name, track name, track start and time played.
}
