% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions_intervals.R
\name{sessions_intervals}
\alias{sessions_intervals}
\title{Sessions length by intervals.}
\usage{
sessions_intervals(streaming_history, mins)
}
\arguments{
\item{streaming_history}{A data.table containing streaming history, after 
'prepare_streaming_history' was used on it.}

\item{mins}{Number of minutes which determine distance between listenings.}
}
\description{
Shows table with sessions divided into listening intervals.
Sessions length are divided into categories: <10min, 10-30min, 0.5-1h, 1-2h, 2-4h, 4-6h, 6h+,
}
