% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions_length.R
\name{sessions_length}
\alias{sessions_length}
\title{Sessions length}
\usage{
sessions_length(streaming_history, mins)
}
\arguments{
\item{streaming_history}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it.}

\item{mins}{Number of minutes which determine distance between listenings.}
}
\value{
A data table containg infromation about length of every session.
}
\description{
Infromation about length of every session. Excludes skipped tracks.
}
