% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_hist_with_playlists.R
\name{str_his_with_playlists}
\alias{str_his_with_playlists}
\title{Prepares the streaming history with playlists data table}
\usage{
str_his_with_playlists(playlists_dt, streaming_history)
}
\arguments{
\item{playlists_dt}{A data table containing the names of playlists and songs on them from spotify.}

\item{streaming_history}{A raw data table containing streaming history from spotify.}
}
\value{
A data table containg streaming history from spotify with information whether song is on user playlist. If it is not, the relevant information will appear.
}
\description{
Prepares the streaming history with playlists data table
}
