% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_hist_with_playlists_wide.R
\name{str_his_with_playlists_wide}
\alias{str_his_with_playlists_wide}
\title{Prepares the streaming history with playlists data table}
\usage{
str_his_with_playlists_wide(playlists_dt, streaming_history)
}
\arguments{
\item{playlists_dt}{A data table containing the names of playlists and songs on them from spotify.}

\item{streaming_history}{A data table containing streaming history from spotify.}
}
\value{
A data table containing streaming history from spotify with information on which 
playlists song is. Data table is in wide format.
}
\description{
Prepares the streaming history with playlists data table
}
