% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_spotifyvis.R
\name{theme_spotifyvis}
\alias{theme_spotifyvis}
\title{Spotifyvis theme for ggplot2}
\usage{
theme_spotifyvis(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family.}

\item{base_line_size}{Base size for line elements}

\item{base_rect_size}{Base size for rect elements}
}
\description{
Theme used in plots.
}
