% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_spotifyviz.R
\name{theme_spotifyviz}
\alias{theme_spotifyviz}
\title{Spotifyviz theme for ggplot2}
\usage{
theme_spotifyviz(
  base_size = 16,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family.}

\item{base_line_size}{Base size for line elements}

\item{base_rect_size}{Base size for rect elements}
}
\description{
Theme used in plots.
}
