% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_search_queries.R
\name{filter_search_queries}
\alias{filter_search_queries}
\title{Filter spotify search queries}
\usage{
filter_search_queries(search_queries, start_date, end_date)
}
\arguments{
\item{search_queries}{A data table with search queries, made with \code{\link{make_search_queries_dt}}}

\item{start_date}{A POSIXt,Date or string that can be coerced into Date by \code{\link{as_date}} indicating start of the period of time.}

\item{end_date}{A POSIXt, Date or string that can be coerced into Date by \code{\link{as_date}} indicating end of the period of time.}
}
\value{
A search queries data table containing only entries between \code{start_date} and \code{end_date}
}
\description{
Filter spotify search queries data table to contain only entries from specified time period
}
\seealso{
\code{\link{filter_streaming_history}}
}
