% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_streaming_history.R
\name{filter_streaming_history}
\alias{filter_streaming_history}
\title{Filter streaming history}
\usage{
filter_streaming_history(streaming_history, start_date, end_date)
}
\arguments{
\item{streaming_history}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it.}

\item{start_date}{A POSIXt,Date or string that can be coerced  into Date indicating start of the period of time.}

\item{end_date}{A POSIXt, Date or string that can be coerced into Date indicating end of the period of time.}
}
\description{
Filer streaming history data table to only contain information about a given period of time.
}
