% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_long_listened.R
\name{how_long_listened}
\alias{how_long_listened}
\title{Measure time spent listening to spotify}
\usage{
how_long_listened(
  streaming_history,
  start_date,
  end_date,
  as_percentage = FALSE
)
}
\arguments{
\item{streaming_history}{A data.table containing streaming history, after 'prepare_streaming_history' was used on it}

\item{start_date}{A POSIXt,Date or string that can be coerced into Date by \code{\link{as_date}} indicating start of the period of time to count tracks}

\item{end_date}{A POSIXt,Date or string that can be coerced into Date by \code{\link{as_date}} indicating end of the period of time to count tracks}

\item{as_percentage}{A logical scalar. If FALSE (default) length of time (in seconds and an approximate in biggest reasonable unit)
that songs were listened to in given time period is returned,
otherwise a character vector indicating percentage of given time period will be returned.}
}
\value{
An ineteger if `as_percetage` is FALSE and a character vector if `as_percentage` is TRUE
}
\description{
Measures time spent listening to spotify during given time period. 
Either as a duration or as a percentage of the whole given period of time.
}
